+++ src/Products/MeetingPROVHainaut/testing.py
@@ -24,5 +24,21 @@ MPH_TESTING_PROFILE = MCLayer(
     gs_profile_id='Products.MeetingPROVHainaut:testing',
     name="MPH_TESTING_PROFILE")
 
+MPH_FIN_TESTING_PROFILE = MCLayer(
+    zcml_filename="testing.zcml",
+    zcml_package=Products.MeetingPROVHainaut,
+    additional_z2_products=('imio.dashboard',
+                            'Products.PloneMeeting',
+                            'Products.MeetingCommunes',
+                            'Products.MeetingPROVHainaut',
+                            'Products.CMFPlacefulWorkflow',
+                            'Products.PasswordStrength'),
+    gs_profile_id='Products.MeetingPROVHainaut:zprovhainaut',
+    name="MPH_FIN_TESTING_PROFILE")
+
+
 MPH_TESTING_PROFILE_FUNCTIONAL = FunctionalTesting(
     bases=(MPH_TESTING_PROFILE,), name="MPH_TESTING_PROFILE_FUNCTIONAL")
+
+MPH_FIN_TESTING_PROFILE_FUNCTIONAL = FunctionalTesting(
+    bases=(MPH_FIN_TESTING_PROFILE,), name="MPH_FIN_TESTING_PROFILE_FUNCTIONAL")